Fix minor bug in Car class
Previously, when calling the default constructor of the Car class, a Car object will be instantiated if both the model name and license plate number are string type values, and the price is a double type value. This means that even negative values, as long as they are of double type, will still be considered as valid and stored as the price attribute of the Car object.

However, in our context of price, negative price doesn't make sense. Hence, the price value entered by the user will first pass through a simple conditional check, and an exception will be thrown if the price is negative.