Add new car exception to handle unknown license plate number when adding a new transaction
When adding a new transaction, the license plate number must be an existing one. (i.e. car and license plate number must exist in the list) If is the license plate number is not found/unknown, it simply means that the car does not exist in the list of available cars yet, hence throwing an exception.