Add new method to format car price to 2dp
Whenever money is concerned and there is a decimal point, the result should always be formatted to 2 d.p. However, when storing the result as a double, the program always removes the trailing zeroes.

For example, when assigning a value of "16.10" to a double type variable "result", the program will store it as "16.1". Therefore, this new method ensures that when the car list is printed out, the price will be formatted to 2dp for ALL cars. (i.e. 16.10 instead of 16.1). In addition, for whole number/integers, only the int value will be printed out. (i.e. 16 if the value stored is "16" or "16.00", etc.)