Update 'isValidLicensePlateNumber' method in 'CarParser'
The updated method improves the flexibility of the license plate number requirement. Previously, the requirement was that the numeric part of the license plate number MUST BE a 4 digit number. Technically, since it has to be a 4 digit number, the accepted range is only from 1000 to 9999. However, in order to bypass this and input a number less than 1000, users can input leading zeroes at the front of the number (e.g. 0055), and this is undesirable.

Based on the real world, the accpetable range of numbers is from 1 to 9999. Therefore, in an attempt to model after the real world and be as realistic as possible, this new updated method allows users to do just that. The updated method now accepts numbers ranging from 1 to 9999, without any leading zeroes.