Update 'formatPriceToTwoDp' method in 'CarList'
Previously, only price with a decimal point will be formatted to 2 d.p. In the event that the specified price is an integer/whole number, the formatted price will still remain as an integer/whole number. In the updated method, ALL prices will now be formatted to 2 d.p, regardless whether there is a decimal point. For example, even if the specified price is 16, the method will return a formatted price of 16.00.